fix: delete file api to check file ownership first

+++ app/api/uploadthing/delete/route.ts
@@ -1,8 +1,23 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { utapi } from "@/server/uploadthing";
+import { db } from "@/db";
 
 export async function DELETE(request: Request) {
   const res = await request.json();
+
+  const fileDBEntry = await db.file.findFirst({
+    where: {
+      userId: res.userId,
+      key: res.key,
+    },
+  });
+
+  if (!fileDBEntry)
+    return new Response("Unauthorized access detected", {
+      status: 401,
+    });
+
   await utapi.deleteFiles(res.key);
+
   return Response.json({ success: true });
 }
